+++ csrc/moves/queenside-castling.cu
@@ -5,7 +5,7 @@
 
 
 __device__ bool queenside_castle_move(
-    int env,
+    size_t env,
     torch::PackedTensorAccessor32<int , 1 , torch::RestrictPtrTraits> players ,
     torch::PackedTensorAccessor32<int , 2 , torch::RestrictPtrTraits> boards  ,
     torch::PackedTensorAccessor32<int , 2 , torch::RestrictPtrTraits> actions
